requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
cryptg==0.4.0
DateTime==5.5
downloader-cli==0.3.4
ffmpeg==1.4
ffprobe==0.5
frozenlist==1.4.1
fuzzywuzzy==0.18.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.9.0
idna==3.7
instaloader==4.11
Levenshtein==0.25.1
lxml==5.2.2
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
mutagen==1.47.0
nepali-datetime==1.0.8.2
playwright==1.48.0
psutil==5.9.8
pyaes==1.6.1
pyasn1==0.6.0
pybalt==2024.11.16
pycryptodomex==3.20.0
pyee==12.0.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-telegram-bot==21.3
pytube==15.0.0
pytz==2024.1
rapidfuzz==3.9.3
requests==2.32.3
rich==13.7.1
rsa==4.9
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
Telethon==1.36.0
typing_extensions==4.12.2
urllib3==2.2.1
websockets==13.0
yarl==1.9.4
yt-dlp==2024.10.31.232951.dev0
zope.interface==6.4.post2
